+++ aeolus/calc/calculus.py
@@ -1,10 +1,13 @@
 """Generic calculus functions."""
-import iris
-from iris.analysis.calculus import differentiate
+from cf_units import Unit
 
+import iris
+from iris.analysis.calculus import _coord_cos, _curl_differentiate, _curl_regrid, differentiate
 
 import numpy as np
 
+from ..const import get_planet_radius
+from ..exceptions import NotYetImplementedError
 from ..model import um
 
 
@@ -13,6 +16,7 @@ __all__ = (
     "d_dy",
     "d_dz",
     "deriv",
+    "div_h",
     "integrate",
 )
 
@@ -59,6 +63,92 @@ def deriv(cube, coord):
     diff = differentiate(cube, coord)
     res = diff.interpolate([(coord, pnts)], iris.analysis.Linear())
     return res
+
+
+def div_h(i_cube, j_cube, r_planet=None, model=um):
+    r"""
+    Calculate horizontal divergence.
+
+    Note: currently works only in spherical coordinates.
+
+    Parameters
+    ----------
+    i_cube:
+        i-component (e.g. u-wind)
+    j_cube:
+        j-component (e.g. v-wind)
+    r_planet: float, optional
+        Radius of the planet (m). If not given, an attempt is made
+        to get it from the cube metadata.
+    model: aeolus.model.Model, optional
+        Model class with relevant variable names.
+
+    Returns
+    -------
+    iris.cube.Cube
+        Cube of horizontal divergence.
+
+    Notes
+    -----
+    Divergence in spherical coordinates is defined as
+
+    .. math::
+
+        \nabla\cdot \vec A = \frac{1}{r cos \theta} (
+        \frac{\partial \vec A_\lambda}{\partial \lambda}
+        + \frac{\partial}{\partial \theta}
+        (\vec A_\theta cos \theta))
+
+    where \lambda is longitude, \theta is latitude.
+    """
+    x_coord = i_cube.coord(model.x)
+    y_coord = i_cube.coord(model.y)
+
+    y_dim = i_cube.coord_dims(y_coord)[0]
+
+    horiz_cs = i_cube.coord_system("CoordSystem")
+
+    # Check for spherical coords
+    spherical_coords = isinstance(
+        horiz_cs, (iris.coord_systems.GeogCS, iris.coord_systems.RotatedGeogCS)
+    )
+    if spherical_coords:
+        # Get the radius of the planet
+        if r_planet is None:
+            r = get_planet_radius(i_cube)
+        else:
+            r = r_planet
+        r_unit = Unit("m")
+
+        lon_coord = x_coord.copy()
+        lat_coord = y_coord.copy()
+        lon_coord.convert_units("radians")
+        lat_coord.convert_units("radians")
+        lat_cos_coord = _coord_cos(lat_coord)
+
+        # j-component: \frac{\partial}{\partial \theta} (\vec A_\theta cos \theta))
+        temp = iris.analysis.maths.multiply(j_cube, lat_cos_coord, y_dim)
+        djcos_dtheta = _curl_differentiate(temp, lat_coord)
+        prototype_diff = djcos_dtheta
+
+        # i-component: \frac{\partial \vec A_\lambda}{\partial \lambda}
+        d_i_cube_dlambda = _curl_differentiate(i_cube, lon_coord)
+        d_i_cube_dlambda = _curl_regrid(d_i_cube_dlambda, prototype_diff)
+        new_lat_coord = d_i_cube_dlambda.coord(model.y)
+        new_lat_cos_coord = _coord_cos(new_lat_coord)
+        lat_dim = d_i_cube_dlambda.coord_dims(new_lat_coord)[0]
+
+        # Sum and divide
+        div = iris.analysis.maths.divide(
+            iris.analysis.maths.add(d_i_cube_dlambda, djcos_dtheta),
+            r * new_lat_cos_coord,
+            dim=lat_dim,
+        )
+        div.units /= r_unit
+        div = div.regrid(i_cube, iris.analysis.Linear())
+    else:
+        raise NotYetImplementedError("Non-spherical coordinates are not implemented yet.")
+    return div
 
 
 def integrate(cube, coord):